+++ packages/core-sdk/src/marketplaces/opensea.ts
@@ -291,7 +291,8 @@ export class OpenSeaMarketplace extends Marketplace {
       fulfillerAddress,
       osOrder.orderHash,
       osOrder.protocolAddress,
-      osOrder.side
+      // osOrder.side
+      OrderSide.LISTING // TODO: Hack waiting for Opensea API bug to be fixed. See https://github.com/fermionprotocol/ui/issues/358
     );
     const inputData = ffd.fulfillment_data.transaction
       .input_data as unknown as {
@@ -473,6 +474,7 @@ export class OpenSeaMarketplace extends Marketplace {
       },
       accountAddress: listing.offerer,
       startAmount: formatUnits(listing.price, listing.exchangeToken.decimals),
+      listingTime: listing.listingTime,
       expirationTime: listing.expirationTime,
       paymentTokenAddress: listing.exchangeToken.address,
       englishAuction: listing.auction