+++ packages/react-kit/src/components/modal/components/Redeem/MyItems/Exchange.tsx
@@ -173,7 +173,7 @@ export default function Exchange({
           }
         }}
         onAvatarNameClick={handleOnAvatarClick}
-        price={Number(price)}
+        price={price}
         currency={offer.exchangeToken.symbol as Currencies}
         {...createSpecificCardConfig()}
       />