+++ packages/react-kit/src/components/productCard/ProductCard.tsx
@@ -33,7 +33,7 @@ interface IProductCard {
   isHoverDisabled?: boolean;
   onAvatarNameClick?: () => void;
   onCardClick?: (id: string | number) => void;
-  price: number;
+  price: string;
   productId: string;
   productType?: ProductType;
   title: string;