+++ packages/react-kit/src/hooks/connection/connection.ts
@@ -67,18 +67,28 @@ export function useChainId(): number | undefined {
   const {
     withExternalConnectionProps,
     externalConnectedChainId,
-    withWeb3React
+    withWeb3React,
+    withMagicLink
   } = useConfigContext();
   const externalSigner = useExternalSigner();
   const { data: externalSignerChainId } = useExternalSignerChainId();
-  const magicChainId = useMagicChainId();
   const isMagicLoggedIn = useIsMagicLoggedIn();
   let networkChainId: number | undefined;
+  let magicChainId: number | undefined;
   let error: unknown;
+  let magicError: unknown;
   try {
     networkChainId = useNetwork().chain?.id;
   } catch (wagmiError) {
     error = wagmiError; // error if the provider is not there
+  }
+  try {
+    magicChainId = useMagicChainId();
+  } catch (error) {
+    magicError = error; // error if the provider is not there
+  }
+  if (!withExternalConnectionProps && magicError && withMagicLink) {
+    throw magicError;
   }
   if (externalConnectedChainId) {
     return externalConnectedChainId;
@@ -86,7 +96,7 @@ export function useChainId(): number | undefined {
   if (externalSigner) {
     return externalSignerChainId;
   }
-  if (isMagicLoggedIn) {
+  if (isMagicLoggedIn && withMagicLink) {
     return magicChainId;
   }
   if (web3ReactChainId) {
@@ -108,32 +118,43 @@ export function useIsConnectedToWrongChain(): boolean {
 export function useProvider():
   | providers.JsonRpcProvider
   | providers.FallbackProvider {
-  const { withExternalConnectionProps, withWeb3React } = useConfigContext();
+  const { withExternalConnectionProps, withWeb3React, withMagicLink } =
+    useConfigContext();
   const { provider: web3Provider } = useWeb3ReactWrapper() || {};
-  let provider;
+  let provider, magicProvider;
   let error: unknown;
   try {
     provider = useEthersProvider();
   } catch (wagmiError) {
     error = wagmiError; // error if the provider is not there
   }
-  const magicProvider = useMagicProvider();
-  const isMagicLoggedIn = useIsMagicLoggedIn();
   if (!withExternalConnectionProps && error && !withWeb3React) {
     throw error;
   }
+  try {
+    magicProvider = useMagicProvider();
+  } catch (magicError) {
+    error = magicError; // error if the provider is not there
+  }
+  if (!withExternalConnectionProps && error && withMagicLink) {
+    throw error;
+  }
+  // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
+  const magicLinkProvider = magicProvider!; // it should always be not null at this point
+  const isMagicLoggedIn = useIsMagicLoggedIn();
   return isMagicLoggedIn
-    ? magicProvider ?? web3Provider ?? provider
+    ? magicLinkProvider ?? web3Provider ?? provider
     : withWeb3React
-      ? web3Provider ?? magicProvider ?? provider
-      : provider ?? magicProvider ?? web3Provider;
+      ? web3Provider ?? magicLinkProvider ?? provider
+      : provider ?? magicLinkProvider ?? web3Provider;
 }
 
 export function useSigner(): Signer | undefined {
   const {
     withExternalConnectionProps,
     externalConnectedSigner,
-    withWeb3React
+    withWeb3React,
+    withMagicLink
   } = useConfigContext();
   let wagmiSigner: ReturnType<typeof useEthersSigner>, error: unknown;
   try {
@@ -142,8 +163,14 @@ export function useSigner(): Signer | undefined {
   } catch (wagmiError) {
     error = wagmiError; // error if the provider is not there
   }
+  let magicProvider: providers.Web3Provider | undefined, magicError: unknown;
+
+  try {
+    magicProvider = useMagicProvider();
+  } catch (err) {
+    magicError = err; // error if the provider is not there
+  }
   const { externalSigner } = useExternalSigner() ?? {};
-  const magicProvider = useMagicProvider();
   const isMagicLoggedIn = useIsMagicLoggedIn();
   const { provider: web3Provider } = useWeb3ReactWrapper() || {};
 
@@ -168,6 +195,9 @@ export function useSigner(): Signer | undefined {
   ]);
   if (!withExternalConnectionProps && error && !withWeb3React) {
     throw error;
+  }
+  if (!withExternalConnectionProps && magicError && withMagicLink) {
+    throw magicError;
   }
   return signer;
 }